float to doule

+++ simple-pyspark-app.py
@@ -107,10 +107,10 @@ def process_dim_data(
                 inner_df \
                 .withColumn("state_bottle_retail_item",
                             regexp_replace(col('state_bottle_retail'), "[^0-9.]", "")) \
-                .withColumn("state_bottle_retail", col("state_bottle_retail_item").cast("float")) \
+                .withColumn("state_bottle_retail", col("state_bottle_retail_item").cast("double")) \
                 .withColumn("state_bottle_cost_item",
                             regexp_replace(col('state_bottle_cost'), "[^0-9.]", '')) \
-                .withColumn("state_bottle_cost_item", col("state_bottle_cost_item").cast("float"))        
+                .withColumn("state_bottle_cost_item", col("state_bottle_cost_item").cast("double"))        
             logging.warning(f"Length of dimension {key} is : {inner_df.count()}") 
         else:
             inner_df = df.drop_duplicates(val).select(val)
@@ -140,7 +140,7 @@ def process_dim_data(
     order_fact \
             .withColumn("sale_total",
                         regexp_replace(col("sale"), "[^0-9.]", "")) \
-            .withColumn("sale_total", col("sale_total").cast("float")) \
+            .withColumn("sale_total", col("sale_total").cast("double")) \
             .drop("sale")
 
     order_fact.write.parquet(f"s3a://{s3_bucket}/{s3_key}/order_fact.parquet",mode='overwrite')
